+++ api/DHSC.FingertipsNext.Modules.Core.Service/IIndicatorsService.cs
@@ -5,6 +5,6 @@ namespace DHSC.FingertipsNext.Modules.Core.Service
 {
     public interface IIndicatorsService
     {
-        IEnumerable<HealthMeasure> GetIndicatorData(int indicatorId, string []areaCodes, int[]years);
+        IEnumerable<HealthMeasure> GetIndicatorData(int indicatorId, string []areaCodes, short[]years);
     }
 }
